pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "defaultpython"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[project.scripts]
defaultpython = "defaultpython.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/defaultpython"]

[tool.hatch.build.targets.wheel.force-include]
"src/defaultpython/py.typed" = "defaultpython/py.typed"

[tool.hatch.build.targets.sdist]
include = ["src/defaultpython"]

[tool.coverage.run]
source = ["src"]
branch = true
relative_files = true

[tool.coverage.report]
show_missing = true
fail_under = 80

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true

[tool.ruff]
line-length = 88
target-version = "py313"
lint.select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG", "ANN"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "bandit>=1.9.2",
    "detect-secrets>=1.5.0",
    "pip-audit>=2.7.3",
    "vulture>=2.11",
    "commitizen>=4.11.3",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.11",
]
